move url mappings to src folder
