+++ .npmignore
@@ -12,8 +12,7 @@ js/tests
 /schema
 package.json
 src
-testdata/**/*
-!testdata/**/shlTestResponses.js
+testdata
 tsconfig.json
 !js/src/api.d.ts
 !js/package.json
